chore: refactor root layout
add sitemap
