+++ public/sitemap.xml
@@ -1,0 +1,14 @@
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
+    <url>
+        <loc>https://notflox.vercel.app/auth</loc>
+        <lastmod>2024-01-22T17:41:26+00:00</lastmod>
+        <priority>1.00</priority>
+    </url>
+    <url>
+        <loc>https://notflox.vercel.app/profiles</loc>
+        <lastmod>2024-01-22T17:41:26+00:00</lastmod>
+        <priority>0.80</priority>
+    </url>
+</urlset>
