+++ src/components/layouts/footer.tsx
@@ -1,29 +1,29 @@
-import Link from 'next/link';
-import { Netflix } from '../icons';
+import Link from "next/link";
+import { Netflix } from "../icons";
 
-const date = new Date().getFullYear();
-
-export default function Footer() {
+export const Footer = () => {
+  const date = new Date().getFullYear();
   return (
     <footer className="bg-ireng transition-colors">
-      <hr className="md:mb-6 pt-4 sm:mx-auto border-gray-700" />
-      <div className="p-2.5 md:p-4 max-w-7xl mx-auto relative">
-        <div className="flex flex-col justify-center mb-4 sm:flex-row sm:justify-between">
-          <Netflix />
+      <hr className="border-gray-700 pt-4 sm:mx-auto md:mb-6" />
+      <div className="relative mx-auto max-w-7xl p-2.5 md:p-4">
+        <div className="mb-4 flex flex-col justify-center sm:flex-row sm:justify-between">
+          <Netflix className="-ml-2" />
         </div>
-        <div className="flex gap-6 flex-row justify-between text-gray-400 items-center">
+        <div className="flex flex-row items-center justify-between gap-6 text-gray-400">
           <span className="flex gap-3 text-sm font-semibold">
             © {date} yogyy.
           </span>
           <Link
-            className="hover:text-gray-600 dark:hover:text-gray-200 font-semibold"
+            className="font-semibold hover:text-gray-600 dark:hover:text-gray-200"
             href="https://github.com/yogyy/Netflix-Movie-App"
             title="source code"
-            target="_blank">
+            target="_blank"
+          >
             Github
           </Link>
         </div>
       </div>
     </footer>
   );
-}
+};
